src/requirements.txt
pyserial==3.5
paho-mqtt
PyYAML